update compresser and uncompresser

+++ core/filecompresser.cpp
@@ -5,39 +5,51 @@
 #include <QDir>
 #include <QProcess>
 #include <QDebug>
-#include <QTimer>
+QString CompressParam::volumeText() const
+{
+    if(volumeSize == 0)
+        return QString();
+    return QString("-v%1m").arg(volumeSize);
+}
+
+QString CompressParam::passwordText() const
+{
+    if(password.isEmpty())
+        return password;
+    return QString("-p%1").arg(password);
+}
 
 FileCompresser::FileCompresser(QObject *parent)
     : QObject(parent)
     , process(new QProcess(this))
-    , compressCount(0)
+    , currentIndex(0)
 {
     connect(process, &QProcess::readyReadStandardOutput, this, [=](){
         while(process->canReadLine())
         {
-            QString text = QString::fromUtf8(process->readLine());
-            qDebug() << text;
+            QString text = QString::fromUtf8(process->readLine()).remove("\r\n");
+            if(!text.isEmpty())
+                emit progress(text);
         }
     });
 
     connect(process, QOverload<int, QProcess::ExitStatus>::of(&QProcess::finished), this,
         [=](int exitCode, QProcess::ExitStatus exitStatus) {
         if(exitStatus != QProcess::ExitStatus::NormalExit)
-            emit error("Program is crash exit");
-        else
-        {
-            if(exitCode == 0 && --compressCount == 0)
-                emit finished();
-            else
-                emit error(errorToText(exitCode));
+            onError("Program is crash exit");
+        else {
+            if(exitCode != 0)
+                onError(errorToText(exitCode));
+            else {
+                if(++currentIndex == argsList.size())
+                    emit finished();
+                else {
+                    process->setArguments(nextArgs());
+                    process->start();
+                }
+            }
         }
     });
-}
-
-void FileCompresser::onDataReady()
-{
-    QString text = QString::fromUtf8(process->readAllStandardOutput());
-    qDebug() << text;
 }
 
 bool FileCompresser::compress(QStringList const& fileNames,
@@ -45,27 +57,21 @@ bool FileCompresser::compress(QStringList const& fileNames,
               QString const& targetFilePath)
 {
     QString app = Utils::compressApp();
+    process->setProgram(app);
     if(param.isSignle)
     {
-        compressCount = fileNames.size();
         foreach(auto const& fileName, fileNames)
         {
-            QString srcFileName;
-
-            if(QFileInfo(fileName).isDir() && !param.isWithPath)
-                srcFileName = fileName + "/*";
-            else
-                srcFileName = fileName;
+            QStringList args;
             QFileInfo fileInfo(targetFilePath);
-            QString dstFileName = fileInfo.dir()
+            QString targetFileName = fileInfo.dir()
                     .filePath(QFileInfo(fileName).baseName() + "." + fileInfo.suffix());
-
-            QString params = QString("a %1 %2")
-                    .arg(dstFileName, srcFileName);
-            if(param.isMoveFile)
-                params += " -sdel";
-
-            process->start(app + " " + params);
+            args << "a" << targetFileName
+                 << getFileNames(fileName, param);
+            setArgs(args, param);
+            argsList << args;
+            qDebug()<< args.join(" ");
+            targetFileNames << targetFileName;
         }
     }
     else
@@ -74,35 +80,78 @@ bool FileCompresser::compress(QStringList const& fileNames,
         args << "a" << targetFilePath;
         foreach(auto const& fileName, fileNames)
         {
-            if(QFileInfo(fileName).isDir() && !param.isWithPath)
-                args << (fileName + "/*");
-            else
-                args << fileName;
+            args << getFileNames(fileName, param);
         }
-        if(param.isMoveFile)
-            args << "-sdel";
-        compressCount = 1;
-        process->setProgram(app);
-        process->setArguments(args);
-        QTimer::singleShot(0, this, [=](){
-            process->start();
-        });
+        setArgs(args, param);
+        argsList << args;
+        qDebug()<< args.join(" ");
+        targetFileNames << targetFilePath;
     }
+    currentIndex = 0;
+    process->setArguments(nextArgs());
+    process->start();
     return true;
 }
 
 void FileCompresser::cancel()
 {
+    if (process->state() == QProcess::NotRunning)
+    {
+        emit finished();
+        return;
+    }
+    process->disconnect();
     process->kill();
+    process->waitForFinished(1000);
+    onError("User stopped the process");
+}
+
+void FileCompresser::setArgs(QStringList & args, CompressParam const& param)
+{
+    if(param.isRecursively)
+        args << "-r";
+    else
+        args << "-r-";
+    if(param.isMultiVolume)
+        args << param.volumeText();
+    if(param.isMoveFile)
+        args <<  "-sdel";
+    if(param.isCreateSFX)
+        args <<  "-sfx";
+    if(param.isEncryption)
+        args << param.passwordText();
+}
+
+QStringList FileCompresser::getFileNames(QString const&fileName,
+                      CompressParam const& param)
+{
+    QStringList newFileNames;
+    QFileInfo fileInfo(fileName);
+    if(fileInfo.isFile())
+        newFileNames << fileName;
+    else
+    {
+        if(param.isWithPath)
+            newFileNames << fileName;
+        else
+            newFileNames << fileName + "/*";
+
+    }
+    return newFileNames;
+}
+
+QStringList FileCompresser::nextArgs() const
+{
+    return  argsList.at(currentIndex);
+}
+
+void FileCompresser::onError(QString const& errorText)
+{
+    emit error(errorText);
     emit finished();
 }
 
-bool FileCompresser::isFinished() const
-{
-    return process->state() == QProcess::NotRunning;
-}
-
-QString FileCompresser::errorToText(int errorCode)
+QString FileCompresser::errorToText(int errorCode) const
 {
     if(errorCode == 1)
         return QString("Warning");
